move the clients functionality to pdfserve_client
